+++ Apatite/Private/Windows/RenderInterface.cpp
@@ -15,7 +15,7 @@ namespace apatite::detail
 
         using namespace std;
 
-        void Clear(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
+        void Clear(float r, float g, float b, float a)
         {
             glClearColor(r, g, b, a);
             glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
